docker/pytesseract/pyproject.toml
[project]
name = "tesseract-benchmark"
version = "0.1.0"
description = "Tesseract OCR benchmark for resource-limited environments"
requires-python = ">=3.11"
dependencies = [
    "pytesseract==0.3.10",
    "pillow==10.4.0",
    "psutil==6.0.0",
    "pdf2image==1.17.0",
]